Made the assembly CLSCompliant.
Added attribute [assembly: CLSCompliant(true)] and changed type of
public Crc properties from uint to long to satisfy CLS compliance.

+++ SharpCompress/Common/Entry.cs
@@ -8,7 +8,7 @@ namespace SharpCompress.Common
         /// <summary>
         /// The File's 32 bit CRC Hash
         /// </summary>
-        public abstract uint Crc { get; }
+        public abstract long Crc { get; }
 
         /// <summary>
         /// The string key of the file internal to the Archive.
